Fix magic methods
